Rebrand to Solar Explorer with globe homepage and photo route
