+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "solarpunk-photo-comparison",
+  "name": "solar-explorer",
   "version": "0.1.0",
   "private": true,
   "scripts": {
